Move map to Map module
