+++ js/src/index.js
@@ -10,7 +10,7 @@ __webpack_public_path__ = document.querySelector('body').getAttribute('data-base
 require('./jupyter-gmaps.less');
 
 // Export everything from example and the npm package version number.
-export * from './jupyter-gmaps';
+export * from './Map';
 export * from './Toolbar';
 export * from './Figure';
 export * from './Heatmap';
